Zuul references cleaner
Zuul mergers creates a vast number of git references under /refs/zuul
which are never garbage collected.

With hundred of thousands of reference, that makes git fetch operations
very slow since get upload all the reference to Gerrit to synchronize
the Zuul maintained repository.  On one of Wikimedia busy repository
(mediawiki/core) we have 55000 such references and it can take up to 18
seconds for a fetch to complete.  I have seen occurences of a merge
taking 2 minutes to complete.

As such, this tiny script clear out reference for which the commit date
of the pointed commit object is older than 360 days (the default).

It is not perfect since a recent reference can well point to an old
object.  Which will be the case on repositories that are barely active.
In such case the ref will be gone although it has been recently created.

A better way will be to vary Zuul references by using month/day which
will let one easily garbage collect them.  But I am being lazy and that
would not let us clear out references using the current scheme.

Example usage:

 zuul_clear_refs.py --verbose --dry-run --until 90 /srv/zuul/git/project

Would show a list of references point to commit dates older than 90 days
and output a message whenever the script would delete them.

Reference:
 https://bugzilla.wikimedia.org/show_bug.cgi?id=68481

Change-Id: Id4e55f5d571ebd5e8271e516f53f8e05c1f78c1a