Update patch set 1
Patch Set 1:

What I've found so far is that any time onMergeCompleted() is called, the event passed in seems to always have event.merged as True and event.commit as a gitsha... so setUnableToMerge() never gets called on the item, and thus the didMergerSucceed() call is ending up True in formatReport()... as such it never selects the correct "depends on a change" failure message and ends up with the default one instead.

Still hunting for the actual cause of the regression here though.

Patch-set: 1