Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

I think that we should support draft-publish events as triggers, but I do not think that it is correct to dequeue changes on draft-published.  I think that in a situation where Zuul has access to drafts, it would be the opposite of the desired behavior.  And in the case where Zuul does not have access, it is not necessary.

Patch-set: 4
Label: Code-Review=-1