Update patch set 1
Patch Set 1:

(1 comment)

> Should the source and triggers be completely separate...

Perhaps -- though it's hard to imagine a source that is not also a trigger, so we should probably make that a requirement.  I think we can continue to evolve toward that.

> Eg the source: in layout.yaml is a name given to a configuration category in zuul.conf. In that configuration category the driver (ie gerrit) is listed along with credentials. This would allow sources from multiple gerrits.

Yes, I was about to suggest something similar on that change -- your idea here is excellent.  Would you be willing to build that and your change off of this one?

Patch-set: 1
Label: Verified=0