Update patch set 1
Patch Set 1: Code-Review+1

This doesn't seem like it will cause any major issues for our plan[0]. However, I'm only +1 because it will mean we have to make our change to the layout.yaml backwards compatible with this format when we add in connections. My feeling is that we should try and do connections first, but I'm not blocking on that especially if not having this is causing pain.

[0] https://etherpad.openstack.org/p/cviAgfbKV6

Patch-set: 1
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1