Update patch set 1
Patch Set 1:

We have a use case where we do want every reported result (when triggered off of a timer) needs to be in a different email thread.
We can't use ref or any change property other than {change.project} 
because those properties don't exist when triggered off of a timer.

Does adding a timestamp property automatically cause all emails to be in a different thread?  I thought it would just make that property available if one wants to use it.

What about adding this property to the NullChange object instead since this is the object that is used when a change is triggered off of a timer?

Patch-set: 1
Label: Verified=0