Update patch set 1
Patch Set 1: Code-Review+2

It doesn't force it to be used. I do like the idea of adding it to the NullChange, but putting it in the base Changeish object is probably best so that it is available in all of them.

I forgot NullChange is a thing so I am ok with this as is.

Patch-set: 1
Label: Code-Review=+2