Update patch set 2
Abandoned

The reason I wrote this was to let us use a configuration file to determine per project jobs override.  James happily implemented this in https://review.openstack.org/#/c/116125/ by adding to the cloner a --project-branch

So this refactor is moot.  Though it might be restored later on if we want to add more features to the clonemap.

Patch-set: 2
Status: abandoned