Update patch set 5
Patch Set 5: Code-Review-1

(4 comments)

In case the project=branch point to a non existing branch, we would fallback to the hardcoded value 'master' instead of --branch  which can be hard to spot.

Or maybe we can consider a wrong branch to be an error and bails out.

Patch-set: 5
Label: Code-Review=-1