Update patch set 3
Patch Set 3: Code-Review-1

Is it possible to just use `find` for this? If so we can just document a useable find command.

Also, don't you need to run git gc after deleting refs so that the objects are deleted too? <- this is what the -1 is for, the other thing is fine if `find` doesn't work.

Patch-set: 3
Label: Code-Review=-1