Add cache-no-hardlinks argument to zuul-cloner
The cache directory is prefixing with 'file://' which forces in
git-clone to always copy files. This is intended for OpenStack usage.

By default, git has a feature to hardlink files if the source
and destination on the same disk. This saves space and speeds up
the cloning process.

This exposes that feature to users of zuul-cloner.

Change-Id: Ib8796f9dc5ccfeac4a3c9b3d30629c7db0f84739