Add cache-hardlink argument to cloner
The cached repository are force copied by prefixing the path with
'file://' which trigger in git clone the copy which is what is intended
for OpenStack usage.

In some case, one might want to use the default git behavior of creating
hardlink when the source and dest are on the same file system.

Change-Id: Ib8796f9dc5ccfeac4a3c9b3d30629c7db0f84739