Cloner: Implement cache-no-hardlinks argument
The cache directory is prefixing with 'file://' which forces in
git-clone to always copy files. This is intended for OpenStack usage.

By default, git has a feature to hardlink files if the source
and destination on the same disk. This saves space and speeds up
the cloning process.

This exposes that feature to users of zuul-cloner.

Depends-On: Ie32b5ec48da643aa248867c28cee520923c9e607
Change-Id: Ib8796f9dc5ccfeac4a3c9b3d30629c7db0f84739