Update patch set 3
Patch Set 3:

> Is it possible to just use `find` for this? If so we can just document a useable find command.

One can potentially craft a find based oneliner but it is not that easy to find out the age of a reference from the command line.   Maybe using git log to have the commit shown if it is older than X days and if a result is shown delete the reference.

I found it easier to write a tiny python script.

> Also, don't you need to run git gc after deleting refs so that the objects are deleted too? <- this is what the -1 is for, the other thing is fine if `find` doesn't work.

Indeed the script only delete the references, which is merely to reduce the number of references sent by "git remote update".

Zuul creates merge commit in that repo and git commit invokes `git gc --auto`  but that apparently does not remove unreachable objects. On a busy repo:

 $ git fsck --unreachable|wc -l
 Checking object directories: 100% (256/256), done.
 Checking objects: 100% (648107/648107), done.
 Checking connectivity: 574447, done.
 81283
 $

Apparently git gc automatically prune unreferenced loose objects after two weeks. But objects that have been packed are kept around.

I guess we should add document some cron job to:
* clear unreferenced objects (both loose and packed)
* pack references (there is a cronjob defined in puppet)
* repack the repo

As for this script, maybe it can "just" give some instruction on completion?

Patch-set: 3
Label: Verified=0