Update patch set 1
Patch Set 1: Code-Review-1

Changeish objects represent the state of the change in the remote system (ie, gerrit).  They shouldn't have local information like this.  Particularly in recent versions of Zuul, they can now be cached for quite a long time, so this might not even work as anticipated anymore.

It sounds like you want to have a time represending when the change was enqueued into the (periodic/timer?) pipeline.  The Item class represents that.  Perhaps that can be made available to the reporter so that the enqueue time can be used.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1