Allow webapp to filter json results
Add a 'change_filter' GET option to the webapp. This is so when
we start overlaying results in gerrit we don't have to fetch the
entire json blob to find a result. We can use this filter to find
just the change we're interested in.

Don't cache results when the filter is used. If we did we would have
to somehow remove old filter queries from the cache database. This is
possible but if it isn't needed we should avoid it.

Change-Id: I99abb447e1ff7baef3f2ccdf66e9266d2c43e6c8