Update patch set 8
Patch Set 8:

This change was written before the connections changes. I left it in order for two reasons:

1) The triple-o people really want this change and it's been up for review for a while.

2) It has unit tests that will confirm the correct gerrit is posted to when using multiple connections in future patchsets.

If it's an issue I can reorder the patches but I'd prefer not to. Instead I'd prefer to merge this once the rest of the connections patches are ready to merge. Thus nobody will use sshkeys in the layout.yaml but we can keep this work.

Patch-set: 8