add support for individual change status in webapp
For doing something like integrating testing results into gerrit we
need to expose the results for and individual change in an efficient
way. We already have all the precomputed state in the zuul webapp
cache, so just make it so that we can ask for and individual change in
the format: /X,Y.json

If the change is not found, it's still a 404.

Change-Id: Ib8d80530cc99c222226f73046c17ab0bbf6e080b