add support for filtering changes in zuul webapp
This expands the rest API for zuul for selecting a portion of the zuul
data.

The supported urls are:

 - /status: return a complex data structure that represents the entire
   queue / pipeline structure of the system
 - /status.json (backwards compatibility): same as /status
 - /status/change/X,Y: return status just for gerrit change X,Y

In addition, /status(.json) can be filtered by parameters that will
return an extracted subset of changes.

 - project=PROJECTNAME

Which returns the changes relevant to just that project.

In both the individual status, and filtering cases the changes are
returned as a simple array, and not in the pipeline structure.

Tests are added to demonstrate this working, as well as ensure 404 is
correctly returned when invalid urls are provided.

Change-Id: Ib8d80530cc99c222226f73046c17ab0bbf6e080b