Update patch set 3
Patch Set 3: Code-Review-1

(6 comments)

I think the problem here is that the returned json is flattened when filtered. This loses information such as the pipeline or queue a change is in.

I think this is fine when looking up a change since that's on a specific path (although for clarity maybe it should be '/change/X,Y' to make it clear that the structure is different to '/status') but when filtering /status?project=abc it seems odd to return a different structure. This, for example, wouldn't work if wanted the zuul dashboard to send it's filtered query back.

Patch-set: 3
Label: Code-Review=-1