Update patch set 3
Patch Set 3:

The pipeline exists in the change itself, the only missing piece of information is the queue name. That could be added.

If we are filtering by openstack/nova, I'm not entirely convinced the current behavior of keeping all changes in the gate queue because there are nova changes, for instance, is actually incredibly useful. This is about search.

That all being said, this is actually the reason I didn't address the filter case in the original patch. I feel like there is 100% agreement on what /status/change/X,Y will do, any why it's useful. I feel like I don't really understand the filter use case, or why one format vs. another would or would not be useful.

So I'd rather separate these out, land the base status one, and let the filter case be addressed later.

Patch-set: 3