Update patch set 11
Patch Set 11: Code-Review-1

The proposed syntax doesn't seem very intuitive to me (particularly, I keep confusing myself as to whether it is saying that all of the approvals specified in the config must match, or all of the approvals on the change must match).

What about this syntax instead, which at least covers the use case you noted in the documentation.  It's perhaps not as flexible as what you proposed, but it would go a long way while keeping the syntax and code more simple:

  require:
    approval:
      - verified: [-1, -2]
        absent: true

That would say "require that a negative verified value be absent" or read more closely to your way of expressing it: "require all approvals on a change not match verified<0".

Does this work, and do we miss any other major use cases this way?

Patch-set: 11
Label: Code-Review=-1