Update patch set 11
Patch Set 11:

I think that may miss out on a use-case where you may want every comment to have a +1 for example.

The problem I see with using something like 'absent=true' is that it applies to every comment, whereas when absent isn't set the verified applies to only one comment.

What if we renamed the configs as:

require-approval-any => require-approval-any-comment
require-approval-all => require-approval-all-comment

Or simply better document that the requirements are on the comment sets not on the multiple requirements.

Patch-set: 11
Label: Verified=0