Update patch set 1
Patch Set 1:

> Nice catch. Could we get a test that exercises this case?

I attempted to write a test today but could not figure it out.   My idea was to override the MergeClient.submitJob() method to throw the exception then assert the item is not in a pending state.

But the scheduler keeps running ( run() ) catch the exception and my test doesn't see it.

I would need a hand from someone smarter / having better knowledge about the test suite.  I am afraid if I proceed further it will be just huge lost of time on my side :-(

Patch-set: 1