Update patch set 2
Patch Set 2:

Nothing like commuting and having dinner to quickly figure out a solution.

I have made the merger.submitJob to throw an exception which cause the scheduler to eventually stop processing the event.  From there, we just have to assert that the merge state is NEW (previously it would be always set to PENDING), which will let the original submitJob reprocess the merge.

I swear it is working on my local computer with python 2.7 :D

Now, I am not very happy with saving the original method in a variable and overriding it with a function defined inside the function.  There must be a proper python/testr/unittest way to handle it but that is beyond my knowledge :/

Patch-set: 2
Label: Verified=0