Update patch set 1
Patch Set 1:

From git 1.7.8 release notes:

 * When populating a new submodule directory with "git submodule init",
   the $GIT_DIR metainformation directory for submodules is created inside
   $GIT_DIR/modules/<name>/ directory of the superproject and referenced
   via the gitfile mechanism. This is to make it possible to switch
   between commits in the superproject that has and does not have the
   submodule in the tree without re-cloning.

So the .git is no more a clone of the repository but simply a file that contains the path to the repository (under /.git/modules/ of the parent repository).

I have changed the test from isdir()  to exists(). That should fix the python2.6 tests that probably run on slightly outdated boxes providing an a pre 1.7.8 git version.

Patch-set: 1