Update patch set 8
Patch Set 8:

The HTTP fetching prints very ugly errors in the log if it fails (e.g. HTTP requires authentication).

I think that if it is not possible to remove it altogether then there should be option to disable the HTTP fetching. It does not affect any of the core functionality anyway.

Or bare minimum would be to decrease the log level to debug.

Patch-set: 8
Reviewer: Gerrit User 13718 <13718@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0