Update patch set 3
Patch Set 3:

David wrote:
> The replication events are included since 2.9 (not 2.9.1) with [1]; for some reason we don't have a v2.9 tag on the replication plugin.

My bad, I have used git tag --contains which reported 2.9.1 (since there is no tag). PS3 now comments it is in 2.9 so it is all good.

Jukka wrote:
> About heartbeat: I'm receiving those events all the time :)
> It might be some home brewn extension, I will check our Gerrit maintenance.
> But maybe it would be better to whitelist supported events and discard all the rest? What do you think? This would solve the heartbeat problem and similar issues with future protocol extensions.

I guess if there are more custom plugins sending events, that backlist will indeed grow out of control and be hard to keep up to date.

On the other hand we have a whitelist of events we actually care about ( accountfield_from_type ) which logs an error on unhanded events.  When we upgraded Wikimedia Gerrit, the spam of error made me notice that new events were added and add to be supported by Zuul, hence the lines with a comment mentioning the version:

 'reviewer-added': 'reviewer',  # Gerrit 2.5/2.6


So maybe we can just use a whitelist and discards any non matching event with no error but a debug message.  On Gerrit upgrade, one can turn on debug log to verify whether there is anything missing (or read Gerrit documentation about stream-events and plugins).  Might make it hard to catch them though.

Patch-set: 3