Update patch set 3
Patch Set 3:

Antoine, do you mean a separate whitelist or simply continue using accountfield_from_type?

Now the whitelisting is kind of a side effect of setting the account field. I'm not sure if that is the best way.

Also, it would be nice to avoid catching the KeyError every time when we receive a heartbeat. Even if we could afford the performance hit, I believe :)

Patch-set: 3