Gear: bind to a specified address
Along with gear's I884e0ddcaac9c7de35b913dc967a47ee854d0a8a, this is
needed to make sure that the local Gear server listens on a correct
interface -- it should be the same one as we use for contacting it later
on.

The code will silently do the old, wrong thing when an old version of
Gear is installed.

Change-Id: Ia5a0d098f57dcdb3a730c43ab56e55f61127afe8