Process event multiple times.
Each time zuul receives 'patchset-added' event it queries gerrit for
additional data. Sometimes that query returns no data, but that's not
a network issue. For some reason gerrit doesn't includes 'data' section in reply:

    2014-11-20 12:39:48,265 DEBUG gerrit.Gerrit:
      Received data from Gerrit query:
      {u'rowCount': 0, u'runTimeMilliseconds': 1, u'type': u'stats'}

That results in an exception during _handleEvent() function:

    Traceback (most recent call last):
    File ".../zuul/trigger/gerrit.py", line 108, in run
      class Gerrit(object):
    File ".../zuul/trigger/gerrit.py", line 98, in _handleEvent
      def run(self):
    File ".../zuul/trigger/gerrit.py", line 368, in _getChange
      else:
    File ".../zuul/trigger/gerrit.py", line 400, in updateChange

    Exception: Change 1191,1 not found

After that the change being processed is missed by zuul until it is updated by
someone else. As a result no jobs are started by the zuul.

To workaround this I propose to return such events back to the queue, to
handle them several times.

Change-Id: If561f02f0c073c1cfab1804042677a6a626411c2