Process event multiple times.
Recently I faced a situation when zuul fails to query event information from
gerrit after receiving a 'patchset-created' event from that gerrit.
Because of this zuul ignores that event until a comment with 'recheck' is added.

The reason is that the event received from gerrit is lost after first exception.

So I propose to return such events back to the queue, and decrease their TTL
to avoid infinite loops.

Change-Id: If561f02f0c073c1cfab1804042677a6a626411c2