Doc job specification branch: handles refs
The branch: filter of a job specification can be used to match the
branch a change targets as well as the ref of a ref-updated event.

That was introduced by 45865f3e https://review.openstack.org/14096 but
lacked documentation.

Mention job.branch handles refs, based on the documentation of the
Gerrit trigger `ref` parameter.

Change-Id: I66c87d2410959d4477f96579f104a2c36ece2986