Update patch set 3
Patch Set 3:

I've been through the changes in GitPython that happened from 0.3.2 to 0.3.2.1 and found the commit that broke things (ff13922f).

There were calls to finalize_process added in objects/commit.py that cause Commit to raise exceptions if git fails with return code 128 when iterating through output (e.g. iterating through items, getting a list of untracked local files). This is technically a bug in GitPython (the comments say that 128 is the exit code when there was a communication problem - that's not always the case), but we need to deal with it here.

There are only two code paths affected by the breaking changes. Commit.iter_items and Repo.iter_commits. We only use one of these code paths in Zuul - Repo.iter_commits, and we only use that in the tests, which this patch set fixes.

Patch-set: 3
Reviewer: Gerrit User 12100 <12100@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0