Add support for 'connection' concept
This uncovered a bug with test_cloner where the builds were not
necessarily done in the order expected in the tests. To make sure
they are in order wait after each gerrit event before continuing.

This is a large refactor and as small as I could feasibly make it
while keeping the tests working. I'll do the documentation and
touch ups in the next commit to make digesting easier.

Change-Id: Iac5083996a183d1d8a9b6cb8f70836f7c39ee910