Update patch set 1
Patch Set 1: Code-Review-1

I gave it a try and end up with a stracktrace in /zuul/lib/cloner.py when preparing the branches:

 override_zuul_ref = re.sub(self.zuul_branch, indicated_branch, self.zuul_ref)

With a ref-updates, the ZUUL_BRANCH parameter is not set and both zuul_branch and indicated_branch are None. re.sub() ends up raising a TypeError, "first argument must be string or compiled pattern".


What we could do is add a ZUUL_BRANCH to ref-updates events. When Zuul receives a ref-updates event it could:

* git remote update
* prune remote branches
* git branch --contains <ref-name> -r

That would yield the remote (Gerrit) branch containing the reference. On my setup with a tag 1.19.22 of the REL1_19 branch that gives:

 cd /zuul/mediawiki/core
 git branch --contains refs/tags/1.19.22
 REL1_19


Knowing the branch, zuul-cloner will be able to determine the fallback branch just like Change events :)

Patch-set: 1
Label: Code-Review=-1