Update patch set 4
Patch Set 4:

(2 comments)

That's a workaround, I agree, but it's for the problem in gerrit, not in zuul or networking/ because gerrit just can't find the event in its database. We could try to find the cause in gerrit, but if it occurs once, it might occurs again. That's why I think the real solution should be done in zuul.
For now I can only think about more correct way of re-processing events, with a separate queue for failed events or something.

Patch-set: 4
Reviewer: Gerrit User 7562 <7562@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0