Update patch set 1
Patch Set 1: Code-Review-2

See https://review.openstack.org/#/c/13934/ for where this documentation was removed.  I think at this point the variable should be removed as well.  Is there a reason to keep it?

I'm only leaving a -2 to reverse the approval on this since it's in the gate.  I am open to reconsidering.

Patch-set: 1
Label: Code-Review=-2