Update patch set 3
Patch Set 3:

I am not certain what is going on with daemon 1.6.  Apparently one can have 'hidden' versions of things on pypi, and that they may not actually be hidden when using the simple index with pip.  I do not understand this fully.

I have written this change: https://review.openstack.org/145080  which helps illustrate the problem.  I have confirmed it fails with daemon 1.5.5, and works with 1.6 and 1.5.7 (which was released since this patch was uploaded).

I think that we should at least rebase this patch on the test, but I'm still open to what, if any, changes we should make to the requirements regarding this.

David, thank you very much for tracking down and fixing the problem.

Patch-set: 3
Label: Verified=0