Update patch set 3
Patch Set 3:

My current understanding of the hidden releases issue is that they are only effectively hidden from the webui, and that pip should end up installing the numerically greatest version if we leave it unspecified.  So most people should end up with 1.6.1.

David, how did you end up noticing the problem with 1.5.5?

That leaves open the question of whether we should be using 1.5 or 1.6.  This message suggests that the current focus of development is 1.5 and that 1.6 has problems: https://mail.python.org/pipermail/distutils-sig/2014-December/025326.html

So perhaps we should set the specifier to be >=1.5.7, <1.6 ?

Patch-set: 3