Add support for skip-if-only job filters
There was previously no way to skip a job if the contents of a given
change did not require it to be run.  This meant that tempest jobs
would run in response to doc or unit test change.  This change
introduces a 'skip-if-only' job configuration, e.g.

jobs:
  # tempest-dsvm-neutron jobs should be conditionally skipped
  - name: ^.*tempest-dsvm-neutron.*$
    skip-if-only:
      - project: 'openstack/neutron'
        files:
          - '^/COMMIT_MSG$'
          - '^docs/.*$'
          - '^neutron/tests/.*$'

With this configuration, matched jobs will be skipped for a given
openstack/neutron change provided that every file in the change
matches at least one of the specified file regexes.

Change-Id: I9b261f03ec00426160d9396f3a20312e89b624d4