Add support for a skip-if filter on jobs
There was previously no way to skip a job if the contents of a given
patchset did not require it to be run.  This meant that tempest jobs
would run in response to patchsets that would not affect their
execution, like those that only modified documentation or unit tests.
This change introduces a 'skip-if' configuration option, e.g.

jobs:
  - name: tempest-dsvm-neutron-full
    skip-if:
      - project: '^openstack/neutron$'
        all-files-match:
          - '^docs/.*$'
          - '^neutron/tests/.*$'
      - all-files-match:
          - '^.*\.rst$'

With this configuration, the job will be skipped for a given neutron
patchset provided that every file in the change matches at least one of
the specified file regexes.  The job will also be skipped for any
patchset that modified only rst files.

Change-Id: I9b261f03ec00426160d9396f3a20312e89b624d4