Update patch set 1
Patch Set 1:

John: The point of allowing jobs to filter changes is to ensure the opposite of what you are worried about.  If a change only includes files that match the files in the filter in the commit message - i.e. not changes that would impact a tempest job - then the tempest job would not run.  The filtering is intended to be on a job-by-job basis, so a change that didn't run tempest jobs would still run all the other non-tempest jobs.  In the example given, we would save ourselves from having to run ~7 tempest jobs for every patch that only touched the doc or test trees.

Patch-set: 1