Update patch set 5
Patch Set 5: Code-Review+1

Sorry this was messy but now ready for final review / approval.

Here is a summary of the patches history:

PS1 asserted the .git in submodule was a file which is not the case with old versions of git
PS2 did the fix
PS3 incorrectly restored PS1
PS4 is a rebase of PS3
PS5 restores PS2, rebased on tip of master

Patchset 5 should pass tests now. I had it on Wikimedia production for close to four months :]

Patch-set: 5
Label: Code-Review=+1