Update patch set 1
Patch Set 1: Code-Review+1

We recently started being hit by that the issue on the Wikimedia setup.

The root cause is that we reuse workspaces between builds. I suspect that the git lock is left behind when Zuul abort a build on a new patchset, if that occurs while zuul-cloner is in a middle of a git config operation that is screwed :-D

Our bug has some traces details https://phabricator.wikimedia.org/T86730

Patch-set: 1
Label: Code-Review=+1