Add a test for enqueuing complex dependencies
To fully validate that the logic in enqueueChangesAhead and
enqueueChangesBehind is working as intended, create a complex
graph of dependent changes and ensure that even when a change
in the middle is the triggering change, the whole graph is
enqueued in the correct order.

This also adds a missing return statement from the addChange
method.  It does not actually alter the result in this case as
the current code has enough repitition in walking up and down
the tree to eventually enqueue everything, however, it might
be possible for it to produce the wrong result in some cases,
or perhaps do so in a less efficient way.  At any rate, I
believe the added return value is more correct.

Also, this adjusts one of the most annoying log messages, where
a perfectly normal false condition was being logged at error
level.

Change-Id: Ie39f24943d878ae7510736250fb3c43c53649de0