Update patch set 1
Patch Set 1:

Note that I didn't manage to fit this feature within the layoutvalidator without changing to much stuff, hence the tests are a bit wacky.

Though it's more simple that way, and it works correctly with the layout from project-config.

In short, it's my first contribution to openstack-infra, comments are most welcome!

Patch-set: 1
Label: Verified=0