Add support for disabling bad pipelines
If a CI system breaks rather than reporting lots of failure
messages back to gerrit (or a reporter) instead disable the
pipeline. This is done by setting a disable at value for the
window so that if it drops below that then we disable all future
reports until the configuration is reloaded.

Change-Id: If67f79db48c38594187ce498e7008f1f77c8ca03