Don't show jobs for non-live changes
Currently the job listing for changes in the status JSON as well
as the jobs that are checked for completion, and listed in the report
all come from a method that operates on changes, not items.
Alter it to operate on items so it can return the empty list for
non-live items.  This is safe within the queue processor because of
the previous change that explicitly checks for non-live items when
determining whether changes are complete.

Include the liveness attribute in the status JSON.

Use a grey dot for non-live items in the status screen.

Change-Id: I8907108a2f8f43c75b9cabe86e7c519790d58027