Fix independent pipeline CRD
If two git-dependent changes in the same project are added to an
independent pipeline, the first should be enqueued as a single
live item in its own change queue, and the second should be
enqueued as a live item behind a non-live copy of the first with
those two in their own change queue.  However, the current
behavior is that they are enqueued as two live items each in
their own change queue.

Once the first is removed, the queue processor would determine
that a dependency of the second change had failed to merge, and
therefore the second failed tests (as if it were in a dependent
pipeline).  This is because the queue processor looked for the
status of dependent changes throughout the pipeline, so any
failed dependent change would cause it to fail a change.

The queue processor now only looks ahead in a given item's change
queue (rather than the entire pipeline) to determine if a needed
change is missing.

The checks to remove old versions or abandoned changes have been
updated to operate on items and consider whether those items are
live.

On an independent pipeline manager, getChangeQueue dynamically
creates change queues (they are static on dependent pipelines).
However, there were many cases where a dynamic change queue would
have been created but not used.  Significant leaks could occur
over time.  getChangeQueue has been altered to return a context
manager so that if a queue is created but not used, it is
automatically removed.

Change-Id: Ieaa0cdbff59e6b77a11c82876f4fd5cb01fe950b