Add ignore-dependencies option
With the previous CRD changes in effect, the merge-check pipeline
became rather large since it needlessly stacked dependent (both
git-dependent and commit-dependent) changes for each change tested.
Since commit-dependencies should not affect whether a change has
a merge conflict, and git-dependencies are already taken care of,
add an option to allow independent pipelines to ignore dependencies.

We will only set this for the merge-check pipeline in OpenStack.

Change-Id: I553446374ac12aa3e3f2e4bbea7ca8fafba42294