Update patch set 2
Patch Set 2: Code-Review-1

This might be removing them from the status.json but it is also removing them from the RPC API (and thus is a regression). Perhaps we should instead separate out the building of the job information from the RPC vs the web server. Either that or drop the information there?

Patch-set: 2
Label: Code-Review=-1