Update patch set 3
Patch Set 3: Code-Review-1

Yep, sounds good to me (re splitting). This is, however, still a regression so I'm hesitant on +2'ing without a version bump. Unless we know that nobody is using this RPC command we probably don't want to break the api.

Happy to be overruled though as I doubt (m)any are using it.

Patch-set: 3