Update patch set 14
Patch Set 14:

Thanks for the patch. I played around with it a little more to see if I could remember why I didn't use abc's in the first place. There is likely a way to do this with abc, but I'm not sure how. I wanted to make sure any sub-classes weren't declaring any new methods that the base class didn't implement (with the exception of methods starting with an underscore [_]).

Thoughts?

Patch-set: 14
Label: Verified=0