Update patch set 16
Patch Set 16: Code-Review-1

Im not convinced this change is really needed. I think the proper way to do this is if a driver is required to implement a method make that an abc.abstractmethod in the base class, otherwise provide a default implementation. When done this way (which is what the parent patches now do) this change is not needed.

Patch-set: 16
Reviewer: Gerrit User 10035 <10035@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1