Update patch set 16
Patch Set 16:

If you look at future changes (ie the connections one) all reporters (and triggers etc) are loaded dynamically, so there is no opportunity to call a new public method that isn't available in all classes.

This is more or less enforced from the point that if you try to call the method on all reporters it'll clearly fail unless you add it into the abc. As such I'm fine with this. (although as you point out we could still theoretically validate it).

Patch-set: 16
Label: Verified=0