Update patch set 2
Patch Set 2:

Ok after reading docs and code for GitPython I think I have mostly answered my questions.

First the commit message really confused me in the stuff about the second writer. I think the intent was to say that the lock may be held by the first writer even though that writer is gone then any subsequent writers will fail to get the lock. I read it as implying that the lock could not be released by the first writer because the second writer exists. Can we clean that info up in the commit message and the comment in the code?

Also, we should not call the lock file private release method. Instead we should call config_writer.release() which should properly call the private method for us. Leaving my -1 in place due to this.

Patch-set: 2