Update patch set 20
Patch Set 20: Code-Review-1

(9 comments)

Ive got some comments that arent a huge issue. I think youre missing deleting lib/gerrit.py?

If you wanted to break this change up and make it much smaller an easy way would be to make a single patch which is the creation of a connection base class and the moving of lib/gerrit into the tree, then just import zuul.connections.gerrit.Gerrit in lib.gerrit without changing any of the consumers. Not a requirement, just an idea for if stuff like this comes up in the future.

Patch-set: 20
Label: Code-Review=-1