Update patch set 2
Patch Set 2:

I don't know if Zuul-cloner is meant to support acting on directories that other instances of itself are concurrently also working with. I can't imagine a situation in which that is intentional.

Basically this happens when a build is aborted. Sometimes we manually abort, but in most cases this lock conflict stems from Zuul itself. Zuul cancels jobs if newer events render the previous job no longer needed (e.g. the change is closed, or a newer patchset version is uploaded). In that case the lock would be left behind and the next run encounters a half-finished workspace.

Patch-set: 2
Reviewer: Gerrit User 6808 <6808@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0